Rewrite the comment tree so we don't operate on a flattened list.
Instead of flattening the tree and performing operations on the
flattened list, build a proper tree and flatten it when we need to
update the UI. This should make implementing loading more comments much
easier. The tree also wraps the raw JSON response types we receive to
make working with them easier.

TODO:
1. Right now the Node types directly contain the raw response data. Make
them into proper wrappers to hide the implementation details of Reddit's
api.
2. Make the View only work on nodes in the tree instead of using the raw
response types from the Reddit API.